requirements.txt
datasets~=3.2.0
numpy~=1.26.4
PyYAML~=6.0.2
torch~=2.5.1
transformers~=4.48.1
tqdm~=4.67.1
progressbar~=2.5
pandas~=2.2.3
matplotlib~=3.7.5
packaging~=24.2
evaluate~=0.4.3
torchinfo~=1.8.0
accelerate~=1.3.0
sacrebleu~=2.5.1
sentencepiece~=0.2.0
sacremoses~=0.1.1
unbabel-comet~=2.2.4
